requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
typer==0.12.0
rich==13.7.0
anthropic==0.39.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
pyyaml==6.0.2
gitpython==3.1.43
ruff==0.6.0
bandit==1.7.9
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==12.0
